попробуем немного поменять

+++ src/server/game/Battlegrounds/Battleground.cpp
@@ -76,7 +76,7 @@ namespace Trinity
                 data << uint64(target_guid);
                 data << uint32(strlen(text) + 1);
                 data << text;
-                data << uint8 (_source ? _source->GetChatTag() : 0);
+                data << uint16 (_source ? _source->GetChatTag() : 0);
             }
 
             ChatMsg _msgtype;
@@ -109,7 +109,7 @@ namespace Trinity
                 data << uint64(target_guid);
                 data << uint32(strlen(str) + 1);
                 data << str;
-                data << uint8 (_source ? _source->GetChatTag() : uint8(0));
+                data << uint16 (_source ? _source->GetChatTag() : uint8(0));
             }
 
         private:
